requirements.txt
requests>=2.30.0
python-dotenv>=1.0.0
telethon>=1.34.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
prometheus-client>=0.20.0
Flask>=3.0.0
